Internals

DocumenterInterLinks.ExtCrossReferences — Type

Pipeline step to expand all @extref cross-references.

This runs before Documenter.Builder.CrossReferences.

DocumenterInterLinks.InterLinks — Type

Plugin for enabling external links in Documenter.jl.

links = InterLinks(
    "project1" => "https://project1.url/",
    "project2" => "https://project2.url/inventory.file",
    "project3" => (
        "https://project3.url/",
        joinpath(@__DIR__, "src", "interlinks", "inventory.file")
    );
    default_inventory_file="objects.inv"
)

instantiates a plugin object that must be passed as an element of the plugins keyword argument to Documenter.makedocs. This then enables @extref links in the project's documentation to be resolved, see the Documentation for details.

Arguments

The InterLinks plugin receives mappings of project names to the project root URL and inventory locations. Each project names must be an alphanumerical ASCII string. For Julia projects, it should be the name of the package without the .jl suffix, e.g., "Documenter" for Documenter.jl. For Python projects, it should be the name of project's main module.

The root url / inventory location (the value of the mapping), can be given in any of the following forms:

A single string with a URL of the inventory file, e.g.

"sphinx" => "https://www.sphinx-doc.org/en/master/objects.inv"`

The root URL relative which all URIs inside the inventory are taken to be relative is everything up to the final slash in the inventory URL, "https://www.sphinx-doc.org/en/master/" in this case.

A single string with a project root URL, for example,

"sphinx" => "https://www.sphinx-doc.org/en/master/",`

which must end with slash. This looks for the inventory file with the name corresponding to default_inventory_file directly underneath the given root URL.

A tuple of strings, where the first element is the project root URL and all subsequent elements are locations (URLs or local file paths) to an inventory file, e.g.,

"Julia" => (
    "https://docs.julialang.org/en/v1/",
    joinpath(@__DIR__, "src", "interlinks", "Julia.toml")
),
"Documenter" => (
    "https://documenter.juliadocs.org/stable/",
    "https://documenter.juliadocs.org/stable/inventory.toml.gz",
    joinpath(@__DIR__, "src", "interlinks", "Documenter.toml")
)

The first reachable inventory file will be used. This enables, e.g., to define a local inventory file as a fallback in case the online inventory file location is unreachable, as in the last example.

A DocInventories.Inventory instance.

Properties

names: A list of project names

inventories: A dictionary of project names to DocInventories.Inventory instances

rx: a Regex that matches any valid @extref expression that can be resolved.

The InterLinks object also acts as a (read-only) ordered dictionary so that, e.g., links["project1"] returns the DocInventories.Inventory for that project.

Search

Free-form search in a particular inventory is possible with, e.g.,

links["Julia"](search)

See the discussion on search in the DocInventories.Inventory documentation. Such a search returns a list of matching DocInventories.InventoryItem instances.

In addition,

links(search)

allows to search across all inventories. This returns a list of @extref strings that could be used to reference the matching items.

Methods

find_in_interlinks(links, extref) – find the URL for an extref

See also

The InterLinks mapping is deliberately reminiscent of the intersphinx_mapping setting in Sphinx.

DocumenterInterLinks.WriteInventory — Type

Pipeline step to write the objects.inv inventory to the build directory.

This runs after Documenter.Builder.RenderDocument and only if Documenter was set up for HTML output.

DocumenterInterLinks.find_in_interlinks — Method

Find an @extref link in any of the InterLinks inventories.

url = find_in_interlinks(links, extref)

finds extref in links and returns the full URL that resolves the link.

Arguments

links: the [InterLinks] instance to resolve the reference in

extref: a string of the form "@extref [project] [[:domain][:role]:]name"